requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes
antspyx==0.5.4
    # via brainstemx-full
astroid==3.3.9
    # via
    #   brainstemx-full
    #   pylint
attrs==25.3.0
    # via
    #   brainstemx-full
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via
    #   brainstemx-full
    #   streamlit
cachetools==5.5.2
    # via
    #   brainstemx-full
    #   streamlit
certifi==2025.4.26
    # via
    #   brainstemx-full
    #   requests
charset-normalizer==3.4.1
    # via
    #   brainstemx-full
    #   requests
click==8.2.1
    # via
    #   brainstemx-full
    #   streamlit
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   brainstemx-full
    #   click
    #   pylint
contourpy==1.3.2
    # via
    #   brainstemx-full
    #   matplotlib
cycler==0.12.1
    # via
    #   brainstemx-full
    #   matplotlib
dill==0.3.9
    # via
    #   brainstemx-full
    #   pylint
fonttools==4.57.0
    # via
    #   brainstemx-full
    #   matplotlib
gitdb==4.0.12
    # via
    #   brainstemx-full
    #   gitpython
gitpython==3.1.44
    # via
    #   brainstemx-full
    #   streamlit
idna==3.10
    # via
    #   brainstemx-full
    #   requests
imageio==2.37.0
    # via scikit-image
isort==6.0.1
    # via
    #   brainstemx-full
    #   pylint
jinja2==3.1.6
    # via
    #   altair
    #   brainstemx-full
    #   pydeck
joblib==1.4.2
    # via
    #   brainstemx-full
    #   scikit-learn
jsonschema==4.23.0
    # via
    #   altair
    #   brainstemx-full
jsonschema-specifications==2025.4.1
    # via
    #   brainstemx-full
    #   jsonschema
kiwisolver==1.4.8
    # via
    #   brainstemx-full
    #   matplotlib
lazy-loader==0.4
    # via scikit-image
markupsafe==3.0.2
    # via
    #   brainstemx-full
    #   jinja2
matplotlib==3.10.1
    # via
    #   antspyx
    #   brainstemx-full
    #   neuroimagingtools
    #   seaborn
    #   suitpy
mccabe==0.7.0
    # via
    #   brainstemx-full
    #   pylint
narwhals==1.33.0
    # via
    #   altair
    #   brainstemx-full
    #   plotly
networkx==3.5
    # via scikit-image
neuroimagingtools==1.1.4
    # via
    #   brainstemx-full
    #   suitpy
nibabel==5.3.2
    # via
    #   brainstemx-full
    #   neuroimagingtools
    #   suitpy
numpy==2.0.1
    # via
    #   antspyx
    #   brainstemx-full
    #   contourpy
    #   imageio
    #   matplotlib
    #   neuroimagingtools
    #   nibabel
    #   pandas
    #   patsy
    #   pydeck
    #   scikit-image
    #   scikit-learn
    #   scipy
    #   seaborn
    #   statsmodels
    #   streamlit
    #   suitpy
    #   tifffile
packaging==24.2
    # via
    #   altair
    #   brainstemx-full
    #   lazy-loader
    #   matplotlib
    #   nibabel
    #   plotly
    #   scikit-image
    #   statsmodels
    #   streamlit
pandas==2.2.3
    # via
    #   antspyx
    #   brainstemx-full
    #   neuroimagingtools
    #   seaborn
    #   statsmodels
    #   streamlit
pathlib==1.0.1
    # via brainstemx-full
patsy==1.0.1
    # via
    #   brainstemx-full
    #   statsmodels
pillow==11.2.1
    # via
    #   antspyx
    #   brainstemx-full
    #   imageio
    #   matplotlib
    #   scikit-image
    #   streamlit
platformdirs==4.3.7
    # via
    #   brainstemx-full
    #   pylint
plotly==6.1.2
    # via
    #   brainstemx-full
    #   suitpy
protobuf==5.29.5
    # via
    #   brainstemx-full
    #   streamlit
pyarrow==19.0.1
    # via
    #   brainstemx-full
    #   streamlit
pydicom==3.0.1
    # via brainstemx-full
pylint==3.3.6
    # via brainstemx-full
pyparsing==3.2.3
    # via
    #   brainstemx-full
    #   matplotlib
python-dateutil==2.9.0.post0
    # via
    #   brainstemx-full
    #   matplotlib
    #   pandas
pytz==2025.2
    # via
    #   brainstemx-full
    #   pandas
pyyaml==6.0.2
    # via
    #   antspyx
    #   brainstemx-full
referencing==0.36.2
    # via
    #   brainstemx-full
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.4
    # via
    #   antspyx
    #   brainstemx-full
    #   streamlit
    #   suitpy
rpds-py==0.24.0
    # via
    #   brainstemx-full
    #   jsonschema
    #   referencing
scikit-image==0.25.2
    # via brainstemx-full
scikit-learn==1.6.1
    # via brainstemx-full
scipy==1.15.2
    # via
    #   brainstemx-full
    #   scikit-image
    #   scikit-learn
    #   statsmodels
    #   suitpy
seaborn==0.13.2
    # via brainstemx-full
six==1.17.0
    # via
    #   brainstemx-full
    #   python-dateutil
smmap==5.0.2
    # via
    #   brainstemx-full
    #   gitdb
statsmodels==0.14.4
    # via
    #   antspyx
    #   brainstemx-full
streamlit==1.45.1
    # via brainstemx-full
suitpy==1.3.2
    # via brainstemx-full
tenacity==9.1.2
    # via
    #   brainstemx-full
    #   streamlit
threadpoolctl==3.6.0
    # via
    #   brainstemx-full
    #   scikit-learn
tifffile==2025.6.11
    # via scikit-image
toml==0.10.2
    # via
    #   brainstemx-full
    #   streamlit
tomlkit==0.13.2
    # via
    #   brainstemx-full
    #   pylint
tornado==6.5.1
    # via
    #   brainstemx-full
    #   streamlit
typing-extensions==4.13.1
    # via
    #   altair
    #   brainstemx-full
    #   nibabel
    #   referencing
    #   streamlit
tzdata==2025.2
    # via
    #   brainstemx-full
    #   pandas
urllib3==2.5.0
    # via
    #   brainstemx-full
    #   requests
watchdog==6.0.0 ; sys_platform != 'darwin'
    # via
    #   brainstemx-full
    #   streamlit
webcolors==24.11.1
    # via
    #   antspyx
    #   brainstemx-full
